refactor(infra): extract frontend composition helpers from AppHost Program
AppHost Program.cs was 291 lines and the back half was two large
if/else branches wiring up the frontend layer differently for
`aspire run` (Yumney.Gateway project + four MFEs) versus publish
(Dockerfile container + YARP gateway). Each branch was ~45 lines
of resource wiring with the same McpServer__PublicUrl tail.

Extract both into a new FrontendComposition.cs with two extension
methods on IDistributedApplicationBuilder:

- AddRunModeFrontend(options, ...apis, mcpServer, keycloak): MFEs +
  Yumney.Gateway project. The MFE-add closure also moves into a
  private AddMfes helper so the E2E vs dev-shell branch is one
  expression instead of two duplicate paths.

- AddPublishModeFrontend(...apis, mcpServer): Dockerfile container
  + AddYarp gateway with route + SSE-forwarder config + the
  /.well-known/oauth-protected-resource route from PR #804.

The McpServer__PublicUrl wiring stays inside each helper since it
depends on the local gateway endpoint reference and is identical
in shape across both modes.

- Program.cs: 291 -> 208 lines
- FrontendComposition.cs: 131 lines (new)
